ABM y Casos de uso de Usuarios
Se modificaron en las clases los ID con LONG en lugar de INT, luego se hicieron los test de usuarios donde se hicieron las siguientes pruebas
En TestAgregarUsuario se agregan usuarios con validaciones por si ya existe el dni, mail o nombre de usuario...
En TestBorrarUsuarios se borran con id y con nombre de usuario
y por ultimo en TestTraerUsuarios
traemos usuarios por Id, por dni, por nombre de usuario, traemos todos los usuarios y traemos usuarios por ID

+++ Gestion-Tickets/src/datos/Persona.java
@@ -1,7 +1,7 @@
 package datos;
 
 public abstract class Persona {
-	protected int idPersona;
+	protected long idPersona;
 	protected String nombre;
 	protected String apellido;
 	protected int dni;
@@ -19,11 +19,11 @@ public abstract class Persona {
 		this.telefono = telefono;
 	}
 
-	public int getIdPersona() {
+	public long getIdPersona() {
 		return idPersona;
 	}
 
-	protected void setIdPersona(int idPersona) {
+	protected void setIdPersona(long idPersona) {
 		this.idPersona = idPersona;
 	}
